requirements.txt
# Core ML
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Pose Estimation
mediapipe>=0.10.8
opencv-python>=4.8.0

# Object Detection
ultralytics>=8.0.0  # YOLOv8

# Data & Viz
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Video handling
av>=10.0.0

# Config
pyyaml>=6.0

# Gemma fine-tuning (optional - run on GCP)
# transformers>=4.38.0
# peft>=0.8.0
# datasets>=2.16.0
# accelerate>=0.26.0
